// frontend/src/types/Bowler.ts
// Defines the Bowler interface for type safety in the application

export interface Bowler {
  bowlerFirstName: string
  bowlerMiddleInit?: string
  bowlerLastName: string
  teamName: string
  bowlerAddress: string
  bowlerCity: string
  bowlerState: string
  bowlerZip: string
  bowlerPhoneNumber: string
}